Merge from DevCoopBSM

+++ src/main/kotlin/bsm/kiosk/kiosk_v2/domain/paylog/Paylog.kt
@@ -1,7 +1,6 @@
 package bsm.kiosk.kiosk_v2.domain.paylog
 
 import bsm.kiosk.kiosk_v2.domain.user.User
-import com.fasterxml.jackson.core.sym.Name
 import jakarta.persistence.*
 import java.time.LocalDateTime
 
@@ -18,7 +17,7 @@ class Paylog private constructor(
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "pay_num", nullable = false)
-  val payNum: Int = 0
+  val payNum: Long = 0L
   var codeNumber: String = codeNumber
     private set
   var date: LocalDateTime = date
